test: move initConfig to enable mocking
